Stemming WIP, cleaned up some comments
Added stemming as part of text preprocessing, implemented similarly as the clean_statements() function; may combine them at a later them, or may instead integrate stemming functionality into the CountVectorizer, but the stemming function as is should work for both the Naive Bayes classifier and the RNN. Will monitor performance cost of stemming implementation as well as its impact on the efficacy of both models. Also cleaned up some comments.